Renamed to async providers
